Unify security.py and hashers.py into encryption.py
